requirements.txt
streamlit
langchain>=0.2.10
langchain-community
langchain-core
langchain-groq
faiss-cpu
python-dotenv
pypdf
langchain-textsplitters